requirements.txt
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
easyocr==1.5.0
entrypoints==0.4
Flask==2.1.2
greenlet==1.1.2
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
ipykernel==4.8.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
opencv-python==4.2.0.32
opencv-python-headless==4.5.4.60
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
playwright==1.22.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pyee==8.1.0
Pygments==2.12.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.1.2
zipp==3.8.0
